requirements.txt
asgiref==3.9.1
attrs==25.3.0
coverage==7.10.6
Django==5.2.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
future==1.0.0
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
PyJWT==2.10.1
python-dotenv==1.1.1
PyYAML==6.0.2
psycopg2==2.9.10
referencing==0.36.2
rpds-py==0.27.1
serializers==0.2.4
sqlparse==0.5.3
typing_extensions==4.15.0
uritemplate==4.2.0
